#409: removing spec files
